Robot Framework/requirements.txt
robotframework
robotframework-seleniumlibrary
pandas
pyarrow
selenium
lxml
html5lib